pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["model", "data_processing"]

[project]
name = "graphrouter"
version = "0.1.0"
description = "A Graph-based Router for LLM Selections"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets==3.3.2",
    "numpy<2.0.0",
    "pandas==2.2.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "tqdm==4.67.1",
    "transformers==4.49.0",
    "litellm==1.52.10",
    "bert-score",
    "torch>=2.0.0",
    "torch-geometric",
    "sentence-transformers",
    "plotly",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[project.urls]
Homepage = "https://github.com/ulab-uiuc/GraphRouter"
Repository = "https://github.com/ulab-uiuc/GraphRouter"
Documentation = "https://ulab-uiuc.github.io/GraphRouter/"

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
